pyproject.toml
[project]
name = "exa-seacrh"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "exa-py>=1.15.6",
    "openai-agents>=0.3.0",
    "tavily-python>=0.7.12",
]
